+++ app/BeaconBridge/Data/Entities/Submission/Project.cs
@@ -1,4 +1,4 @@
-namespace BeaconBridge.Data.Entities;
+namespace BeaconBridge.Data.Entities.Submission;
 
 public class Project
 {
