+++ app/BeaconBridge/Data/Entities/Submission/ProjectTreDecision.cs
@@ -1,6 +1,6 @@
 using BeaconBridge.Constants.Submission;
 
-namespace BeaconBridge.Data.Entities;
+namespace BeaconBridge.Data.Entities.Submission;
 
 public class ProjectTreDecision
 {
