+++ app/BeaconBridge/Data/Entities/Submission/SubmissionFile.cs
@@ -1,6 +1,6 @@
 using BeaconBridge.Constants.Submission;
 
-namespace BeaconBridge.Data.Entities;
+namespace BeaconBridge.Data.Entities.Submission;
 
 public class SubmissionFile
 {
